+++ test/helpers/udp_connection.exs
@@ -1,0 +1,3 @@
+defmodule Instream.TestHelpers.UDPConnection do
+  use Instream.Connection, otp_app: :instream
+end