+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enhancements
   - Fields can be defined with default values
   - Tags can be defined with default values
+  - Writing can be done over UDP using `Instream.Writer.UDP`
 
 ## v0.7.0 (2015-10-22)